[bugfix] strlen type case

+++ src/core/containers/string.c
@@ -88,5 +88,5 @@ ntt_StringView ntt_stringToView(ntt_String* pStr)
 
 usize ntt_StrLen(const char* cString)
 {
-	return strlen(cString);
+	return (usize)strlen(cString);
 }
